Move registration into Open telemetry extensions.
